feat(backend): add manual email resend endpoint
